+++ lib/comment/widgets/widgets.dart
@@ -5,8 +5,5 @@ export 'comment_bottom_sheet/general_comment_action_bottom_sheet.dart';
 export 'comment_card/comment_card.dart';
 export 'comment_card/comment_card_actions.dart';
 export 'comment_card/comment_card_header.dart';
-export 'comment_card/comment_card_header_date.dart';
-export 'comment_card/comment_card_header_reply_count.dart';
-export 'comment_card/comment_card_header_score.dart';
 export 'comment_card/comment_content.dart';
 export 'comment_card/comment_depth_indicator.dart';
